requirements_logimancer.txt
torch
datasets
openai
transformers
pytorch-lightning
sentencepiece
pandas
tqdm
aima3
nltk
penman
amr-logic-converter
scipy